regression detector: escape message double quotes
The Markdown report contains double quote characters. If these
characters are not escaped, the double quote characters in the report
are interpreted as JSON string delimiters, yielding an ill-formed JSON
payload. This commit substitutes double quotes in the Markdown report
with escaped versions so that the JSON payload for the pull request
commenter service is less ill-formed (ideally, it is well-formed, but
there may be other bugs).

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>