regression detector: move pr comment into new job
The `single-machine-performance-regression_detector` job's output is a
bit untidy due to the operations needed to transform the Regression
Detector report from Markdown into JSON-escaped Markdown. This commit
intends to replace #29683 by building on the work done in #29684,
essentially transplanting the PR-comment-related commands into a
dedicated job in the spirit of #29683. By using the `pr-commenter`
container from Developer Experience, we avoid having to download an
internal auth tool separately.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>