usm: file registry: Propagate activation error
When the activation fails and we add something to the blocked list,
actually report the error from the activation.  Currently it gets
overwritten from the error from the deactivation which is usually nil.

This is only useful when debugging using the debug endpoint or during
testing, since the error value is not used in other cases.