usm: file registry: Fix handling of short-lived processes
If a process terminates before we open the shared library that we want
to hook, then os.ErrNotExist is returned and the code in
file_registry.go::Register does not add the path to the block list.

But if the process terminates after we successfully open the ELF, then
the probe registration fails due to the writes to uprobe_events failing
because the path (/proc/PID/root/...) no longer exists, and the the
probe registration fails with a different error (the eBPF manager does
not join the errors but instead combines the error strings) and the path
ends up getting blocked.

While the eBPF manager could also be fixed to join the errors, we should
make the file manager more robust against failures in activation due to
short-lived processes by checking if the file still exists before adding
the path to the permanent block list.