usm: Do not block file on environment errors
If an attachment failure has to do with something other than the file
itself, such as the mount path or the process name, do not permanently
block the file from being attached later, since the same file may be
used later from an environment where it should not be blocked.

Note that there is a risk that if the file is simultaneously used both
from an environment where it is not supposed to be blocked, and from an
environment where it should be, after this commit it will start being
attached which may cause issues.