usm: pclntab: Fix bug with finding exact symbols
In case we were provided with symbolSetFilter, we didn't consider null byte in our size of funcNameHelper
thus, we would filter out symbols with length equals to the max symbol.

The bug was missed as currently the code works with prefixSymbolFilter, which has high maxLength setting.

We're also introducing a UT to ensure the code won't break again