usm: process monitor: Increase channel size
Increase the channels sizes in the process monitor since there are many
dropped events reported in both local testing and on staging due to
these sizes.  With the new sizes no hits were seen in the debug metrics
on one staging cluster.

The metrics used to tracked it are the following (the latter is not
relevant unless event stream is enabled).

 datadog.system_probe.usm.process.monitor.process_exec_channel_is_full
 datadog.runtime_security.event_monitoring.events.dropped{consumer_id:process_monitor}