Change AllSettings to use a WRLock instead of RLock
We observed flaky test failures caused by AllSettings having a data race. Investigating it lead to the discover that AllSettings seems to be modifying leaf nodes with map values by building the map of all settings. This means the associated methods should be fully locking instead of only locking for reading.