Update file experiments to the use of logrotate FS
This commit updates the file_to_blackhole experiments to rely on the lading
logrotate FS generator, allowing us to assert on whether the log Agent loses
bytes in the course of its operation. I've added a new check asserting that no
bytes are lost, although I do not expect this to pass on higher latency intake
experiments yet.

REF https://github.com/DataDog/datadog-agent/pull/30744
REF https://github.com/DataDog/lading/pull/1090

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>