[smp]: prevent empty regression detector comments
If the Regression Detector fails to submit a job to SMP, an empty
Regression Detector PR comment is posted to PRs. This empty comment is
noise and undesirable, so this commit tests whether the Regression
Detector generated a Markdown report before attempting to post. If the
Markdown report does not exist, the Regression Detector PR comment
will echo an error message and emit a non-zero exit code so that the
PR comment job fails before posting.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>