pkg/fleet/telemetry: explicitly disable appsec from tracer
At least one customer was startled by the presence of the appsec startup
error produced by the tracer when appsec features are explicitly enabled
but the environment or build does not permit the use of it:

```
2024/10/31 18:33:04 Datadog Tracer v1.69.0 ERROR: appsec: threats detection cannot be enabled for the following reasons: go-libddwaf is disabled when cgo is disabled unless you compile with the go build tag `appsec`. It will require libdl.so.2. libpthread.so.0 and libc.so.6 shared libraries at run time on linuxappsec: no security activities will be collected. Please contact support at https://docs.datadoghq.com/help/ for help.
```

This change explicitly disables appsec before starting the tracer in
order to avoid this behavior. It also redirects all tracer logging
output to the agent's own logger, so that all the agent's logging is
centrally controlled.

Ref: SCRS-1135