[APM] Resolve ContainerID on Stat payloads
Stat payloads computed by the tracers, and received on the agent, are
currently lacking the ContainerID, and the associated Container Tags.

These fields are only necessary under certain conditions:
- Depending on `enable_cid_stats` and `disable_cid_stats` FFs
- Whether or not we are on Fargate

This PR takes the payload request HTTP headers, and resolves the
Container ID, in the same way to do it for trace payloads.