Remove all commented out or unused properties from java.security
This file (our customized version of it at least) is not supposed to
be documentation of available properties. Properties that were set are
specific to a Sun/Oracle providers and will not be used by
BouncyCastle.

BouncyCastle does honour disabledAlgorithms properties, but the
default values are specific to Sun/Oracle providers. We don't need to
impose more restrictions on top of what approved_only mode does.