[EBPF] attacher: Add telemetry for procfs retries
There is code to retry procfs reads for up to 10ms after receiving
events "because procfs could be slow to update". But this makes handling
of short-lived processes much slower and leads to the callback channels
getting full and missing events.

It is hard to see how procfs could ever be inaccessible after an exec
event is received (provided the process is still running). Add some
telemetry to check if this case ever hits or if it can be removed.