usm, npm, classification: Fix a leak in connection_protocol
If USM was enabled, but classification feature of NPM was disabled, we had a leak
in connection_protocol map, as both USM and NPM would have mark the connection as visible
from their probes (NPM would have done that from 'update_conn_stats'), but the probe
to clean the entry ('kretprobe__tcp_close_clean_protocols') wasn't enabled.

To clean an entry, we need all modules to mark it as ready for deletion, but only USM
would have done that. The fix it to prevent NPM from processing the protocol if classification
is disabled. Hence, NPM won't mark the conneciton as accessible from it, and we will only need
USM to delete it