[EBPF] attacher: Remove procfs update wait
There is some code to retry procfs reads for up to 10ms after receiving
exec events "because procfs could be slow to update".  This retry loop
makes handling of short-lived processes much slower and leads to the
callback channels getting full and missing events.

The code has been present from the initial version[0] of the process
monitor.  In that initial version the retry loop was waiting for
procutil's NewProcess() to succeed after receiving an exec event from
netlink.  It's hard to see why it was needed even then, since:

 (1) the netlink event is emitted in the kernel after exec is completed
 (2) for exec to even be called, the process has to already have been
     forked and be running
 (3) if the process is running the /proc/PID will exist, there is no
     race in the kernel here.  Such a race would for example imply that
     /proc/self is unreliable.

In the current code, the delay is waiting for /exe and /comm and not the
/proc/PID directory itself, and the exec is emitted by netlink or by the
event stream.  Even in this case it's hard to see what race is being
addressed since the exec event is emitted after the binary has been
determined and the argument page is set up.

Furthermore, recently metrics were added to check if the "temporary
failure followed by success on retry" case ever happens. This has been
running on nightly clusters for some days now and that case has not been
seen[1] even once.

Remove it since it's apparently not needed.

[0] https://github.com/DataDog/datadog-agent/pull/14706/files#diff-2945b330cb2d393879b702461e6be88cc1de711796c00acb8bcbe83694e90afbR143-R159
[1]
https://ddstaging.datadoghq.com/metric/explorer?fromUser=true&state=H4sIAAAAAAAAA12RXYrDMAyE76LHxZSmbbbUV1mKMbGcCvyTWk4hDbn7oqTtwr5Jw-fRSJ7hgYUpJ9Bw2B9OpmnMvgEFfbHDjUH_zODQU6K6QjNUqgFBA6itNExP6Zvvj2ID9WIY0FcRp0GAShEZCyGDgoL3EbluAwrykBOj8blEW_-z93GrBE02ipdIk8R0tlrDeSydyBFroe79ZgIN9tFrnrhi3HXDuBsZy_y1wHJVIMPGYDffV_NnvVyX6yIYD4GqcRQxyZ0E_6hdTp761SBQpAq62SvgXKpcKheHBTQ45A5WL1nLl3WFGbha4Zrz8dCc28uxbc8nBZjcW7u0L62gL8g3E7OTLTmQo9SDgsGOjA60t4FxUeApVCyviJ_I5kFP8_6EPATiCssvbpvrpvkBAAA&start=1731574771705&end=1732179571705&paused=true#N4Ig7glgJg5gpgFxALlAGwIYE8D2BXJVEADxQEYAaELcqyKBAC1pEbghkcLIF8qo4AMwgA7CAgg4RKUAiwAHOChASAtnADOcAE4RNIKtrgBHPJoQaUAbVBGN8qVoD6gnNtUZCKiOq279VKY6epbINiAiGOrKQdpYZAYgUJ4YThr42gDGSsgg6gi6mZaBZnHKGniqyMkIGFA4MAB0GlgaCHCqTvLaOABGcI1wvfKCjZ61mWzajd04mYIajUZ1wABUPAAEvVgbwHZ4aAg8YxpOmfgiCAAUAJQgPAC6VK7uBxih4S+qbzGl8fdPEBtLBoHKgeQYUEIdrKeowM6YDQaCCZRJoURwJxyRTlHDoqBojFOehMZQiNweND3fgQeyYLBYhQ5EDokRKR48PhA+TohAAYSkwhgKBEBzQPCAA