Specifically provide CPU, memory allotments in quality gates
This commit places specific CPU and memory limits on our idle quality gate
experimnets. By default the target has the runner's full allotment of resources
and as we have asserted a bound on memory it makes sense to bound the memory
available to the target to more closely mimic deploy conditions.

REF https://github.com/DataDog/datadog-agent/pull/31326
REF https://github.com/DataDog/datadog-agent/pull/31246

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>