usm: test: Do not use splice in transparent proxy
Supporting splice(2) with alternative kprobe-based data hooks is
complicated since it requires accessing data available via page
references instead of direct pointers.  Disable the use of splice(2) in
the transparent proxy by default to allow an initial version of the
alternative data hooks to be evaluated with protocols/setups which
normally don't use splice(2) but only need to support it due to the use
of the transparent unix proxy in the tests, for example Go kafka clients
taking to Kafka servers.