event monitor: tests: Allow setting cached hostname
When creating spinning the event monitoring, we spend about 5s in each time for trying to fetch
the hostname. Since we're in a test context we don't really care about the host name.
We allow setting the cached hostname to spare the fetching process.

The PR modifies the NewTestProcessConsumer utility to make it automatic