usm: process monitor: Increase number of runners
We used to run X workers, where X is defined as the value returned from runtime.NumVCPU().
However, if GOMAXPROCS is defined, we rely on it. For our helm/k8s environments it is set to 1
as we're limited to less than a single core. We use the correct number of cpus over this setting
to process faster the events