usm: sharedlibraries: Add fexit probes for openat2
Adding fexit version for openat2 probe, to replace (decided on runtime) the 2 existing
tracpoints. Instead of having a tracpoint for 'sys_enter_openat2' and 'sys_exit_openat2',
we can use fexit on openat2, to get both input arguments and return code in a single
probe, rather then caching the input arguments via a tracepoint on 'sys_enter_openat2' and
pops them, and fetch the return code in the tracpoint on 'sys_exit_openat2'.

The change will cut the perofrmance impact on 'open' syscall by ~50%, as we now use
a single probe instead of 2 probes.