usm: sharedlibraries: programtically decide which probes to use
If the kernel supports openat2 syscall, it means that we're atleast on 5.6. Fexit is supported from 5.5.
Thus, if the kernel supports openat2, then we use fexit on it, and ignore the tracpoints for openat and open.
When openat2 syscall exists, the internal implementation of 'open' and 'openat' are folded into openat2, so
we don't loose coverage by the change