Have config generate a Profile
The .Metrics, .MetricTags, etc. fields on CheckConfig are storing all the same
attributes as a profile; it is easier if we just create a Profile from the
CheckConfig and pass that around. This also removes the caching of those
fields from the CheckConfig; a later commit caches this generated profile in
the DeviceCheck.