discovery: e2e: Avoid fail on temporary network errors
The server provisioning has been seen to fail in some rare cases due to
temporary network errors or server issues on external servers. Avoid
failing the test in the case by (1) retrying the provision after a delay
if it fails and (2) marking the test as skipped if it fails a second
time.