Update Regression Detector experiments to have own target allotments
This commit updates the Regression Detector configuration so that every
experiment has its own memory and CPU allotments. We scale these down
where memory checks do not exist to a generous 2GiB and where memory
checks are set we leave a buffer on top of the check limit. CPU is limited
now to 4 across all experiments.

Signed-off-by: Brian L. Troutwine <brian.troutwine@datadoghq.com>