[RC] Support configuring the core RC service with JWT auth
Private Action Runners are now leverging the core RC service in order to
receive remote config updates. They require a different authentication
scheme than Datadog API keys. This allows for the core service to be
configured to use a JWT for authentication to the backend.

It is not configurable via the command line or an agent YAML file, as it
is only used programmatically from within the Datadog Private Action
Runner. There are no other use cases.