[serverless/proxy]: override default ErrorHandler
The default error handler logs "http: proxy error: %v" then returns an HTTP 502 (bad gateway)
response. This is unfortunate because it lacks much any context on the original request that
failed, and the commonly observed error today is "context deadline exceeded", which is not
actionnable if you don't know what request it was for. It also logs to STDERR and does not
honor the agent's log level.

To address this, we introduce a custom error handler function that logs the error with some
relevant dynamic context using `log.Debugf`, and returns an HTTP 502 and HTTP 504 response
depending on whether the error is a `context.DeadlineExceeded` error or not.