usm: Include PathId in debug traced program list
Currently it's possible to see things like the below where the same
paths are displayed as part of separate entries in the traced programs
list with no hint as to why this is the case.  Include the PathId to the
traced programs list to make it clear exactly which file (dev/inode) is
being traced.

Before:

 {
   "FilePath": "/usr/lib64/libssl.so.1.0.2k",
   "PIDs": [
     4275,
     1111324,
     3971,
     4240,
     4416,
     1110985
   ],
   "ProgramType": "shared_libraries"
 },
 {
   "FilePath": "/usr/lib64/libssl.so.1.0.2k",
   "PIDs": [
     7562
   ],
   "ProgramType": "shared_libraries"
 },

After:

 {
   "FilePath": "/usr/lib64/libssl.so.1.0.2k",
    "PathID": {
      "Dev": 64516,
      "Inode": 655961,
    },
   "PIDs": [
     4275,
     1111324,
     3971,
     4240,
     4416,
     1110985
   ],
   "ProgramType": "shared_libraries"
 },
 {
   "FilePath": "/usr/lib64/libssl.so.1.0.2k",
    "PathID": {
      "Dev": 130,
      "Inode": 4421665,
    },
   "PIDs": [
     7562
   ],
   "ProgramType": "shared_libraries"
 },