usm: watcher: Add periodic scan
There are cases where the eBPF shared library watcher can miss hooking
shared libraries in running processes.  For example:

 (1) When a process such as nginx daemonizes itself by forking.
 Depending on timing, the parent process may exit before the hooking is
 complete leading to the hooking of the library failing, and the child
 process will never be noticied by the watcher since it never opens the
 shared library (since it's inherited from the parent).

 (2) When the path length is longer than what the eBPF program supports.

To ensure that we don't miss these kind of cases, introduce a periodic
scan to try to attach to running processes.  We maintain a map to ensure
that we don't scan processes over and over again to limit CPU usage.  We
scan every processes twice in order to avoid races with startup where we
may scan a process' maps file before it has opened its shared libraries.