usm: sowatcher: Added a method to determine if fexit is supported
Added a method to dynamically decide if fexit is supported on the system.
While fexit is supported from 5.5 kernels for x86, and 6.1 for arm, it was
unstable and missing from arm kernels up to 6.4. The best way to determine
the existance of the feature is by trying to load a dummy fexit of the probe
and checks if it works. The method was taken from CWS.