Set Default site when site is empty string in col config and unset in core config
Currently, when site is empty string in col config and unset in core config this will lead to "" being the site. This PR makes it so we set the default "datadoghq.com" in that case.

This is especially necessary because helm chart default for string is an empty string: https://github.com/DataDog/helm-charts/blob/datadog-3.86.0/charts/datadog/templates/_otel_agent_config.yaml#L24