(fleet) fix the version used to reference the installer from the install scripts
Currently we use `RELEASE_VERSION` to reference the version of the installer in the install scripts. This is broken since `RELEASE_VERSION` is `nightly-a7` and not the expected version, even during a stable build on a tag. This PR fixes this by using the actual version generated in `setup_version`.