discovery: module: cache all fields to allow generating events from serviceInfo
This commits adds additional fields to serviceInfo to allow it to be
enough to generate service events. While not strictly needed for start
events, this will become necesary to generate heartbeats and stop
events. The later specifically, can't depend on the process to get some
of the fields, hence the necessity to cache more fields to preserve the
payloads as they are currently.