service discovery: Reduce container status cache time
The GetContainers() call currently uses a cache validity time of 1
minute for the container information (the information which is scraped
from system-probe, not for the information from workloadmeta).  This is
problematic, since due to jitter, even though the agent check is called
once every minute, the time between the calls to GetContainers() is not
guaranteed to be larger than one minute, so the container information
may not get refreshed when needed, leading to temporarily missing
container information for new containers.

For example, in the log below, the second call is about a half a
millisecond shy of one minute:

 now: 2025-01-17 11:01:35.453556101 +0100 CET m=+164.914910536
 now: 2025-01-17 11:02:35.453245896 +0100 CET m=+224.914600324

Furthermore, since there is a per-container ID cache in
collector_cache.go which also caches errors, and another collector-wide
refresh time (in RefreshCgroups() in collector_linux.go), it's possible
for the the cached "unable to find container ID" error to take two calls
to go away, but the potential->running transaction only takes one call
(1 minute) to happen so the start-service event could be emitted without
the container information.

Also, during testing or debugging when the endpoint is accessed directly
as a debugging aid, the problem is aggravated since the intermediate
calls always happen less than 1 minute apart.

Reduce the time to a couple of seconds similar to the code in the
container check.