usm: sharedlibraries: Ensure path does not exceed path->len.
Ensuring we do not send to user mode a path, just because later in the memory we have seen a matching .so file.

For instance, having the following memory '<path 1>\0libssl.so', would have caused us
to send to the user mode '<path 1>', even if it wasn't matching our filtering.

That was caused as we never verified we're not exceeding path->len, so we scanned
the entire 220 characters we read from the memory. And if it contains a matching pattern
the first path would have been considered as a valid match