service discovery: Simplify end-service handling
There is some code to skip sending the end-service event in two cases:

 (1) If there is another service with the same name which is
     either has a start-service or a hearbeat event being sent
     in the same tick.

 (2) If there is another potential service (i.e. a service which
     as not yet run for 1 minute) with the same name as the one
     in the end service event, to handle service restarts.

The reason mentioned in a comment for (1) is so that the "services don't
disappear in the UI" if two (unrelated) processes have the same
generated service name.  However, this reason is invalid since only the
row for the PID being ending would be removed in the backend, and the
other row for the other PID would still exist.  Also, it's not
guaranteed that the other service would even have a start-service or
hearbeat event in the same tick so this check doesn't always work.

For (2), it's better to handle this in the backend where (a) the margin
for the time to wait for a potential restart can be easily controlled
and (b) the logic for considering two reports to be for the same service
can be easily tweaked and improved (for example using more fields than
just the generated name).

(Note also that as of writing the end-service event is currently ignored
in the backend, so no sync is needed with changes there.)

Remove this code to simplify the implementation.