service discovery: Remove repeated name telemetry
There is telemetry which reports an error if we get more than one
service with the same generated name.  This information is obtainable by
querying the tables on the backend, where it's also possible to
optionally combine it with other fields which are used by the backend to
disambiguate the service, so remove it from the agent to simplify the
code.