exclude `.git` from copyright linter (#33359)
When doing the copyright linting we collect all files matching *.go. If a branch is named with a name ending with .go, like NDMII-3313-integration-metric-stats.go you end up with a file .git/logs/refs/remotes/origin/NDMII-3313-integration-metric-stats.go that we are trying to lint. This PR exclude the .git folder from the linter.