Set SO_PASSCRED before binding the socket
Linux kernel only attaches process credentials if the socket is
unconnected or when either has requested it. Setting SO_PASSCRED on a
socket already bound then races with the data being sent by the
dogstatsd client.

Using net.ListenConfig#Control callback allows us to modify the socket
state before the socket is attached to an address, preventing the
clients sending data without SO_PASSCRED being enabled on a socket and
eliminating the race.