Change bininspect.InspectWithDWARF to take dwarf data
This moves the loading of DWARF data into memory from inside the
bininspect InspectWithDWARF API to outside of it. This allows
dynamic instrumentation to re-use the allocated DWARF data which
would then get passed in. This should drastically reduce the size
of allocations in Go DI while having no effect in the one other
place where the function is used.

Signed-off-by: grantseltzer <grantseltzer@gmail.com>