usm: http2: tests: Function to generate huffman encoded string with a given size
The commit adds a method to generate a random string such that its huffman encoded length is known upfront.
Since there's no way to know what will be the length of the huffman representation of a given string without
encoding it, we generate random strings until we find a valid input.