fix(ssi): standardize auto instrumentation config
This commit standardizes the auto instrumentation config so that
everything uses the same root object. In an ideal world, each component
would only use it's own config. However, this package historically was
configured with a raw datadogConfig object and the options come from all
over the place. This is a good first step to cleaning that up while also
being able to support parsing/validating the config once and using
structs to configure components.