(fleet) migrate to the new configs schema with broader support
Due to backend restriction the existing schema could not work in some scenarios with integrations. This migrates from a schema where config paths were set using JSON keys to a schema where config paths are in a nested field "path".

I added a layer to remap old configurations to this new format transparently.