service discovery: Truncate Java classpath arguments
The Java classpath command line arguments can be very large (thousands
of characters in some cases) and lead to more inmportant information
such as the name of the class being executed to get left out due to
command line size restrictions.  Truncate the entries if we see that we
can't fit the whole command line.