service discovery: Remove invalid assertion in TestPorts
Do not assert that no client ports are present in the list of ports
received from the endpoint, since this check can spuriously fail since
the test infrastructure opens a listening TCP socket on an ephimeral
port, and since we mix the different protocols in one list (TCPv4/v5,
UDPv4/v6) we could happen to find that on the unexpected port list.

To keep the assertion, we would have to change the endpoint to return
information about the protocols also, so that we can make an accurate
comparison. For now, just demote the assertion to a log.