usm: protocols: Don't exclude probes, maps and TCs if those are used by a supported protocol
We can have shared maps, probes and TCs between supported and unsupported modules.
Today, we would just disable those resources, preventing from sharing resources.
After the fix, we are sensitive to this scenario and allowing it.

It will be crucial for the separation of istio and nodejs protocols from the native tls
implementations, as we share probes and maps between them.