Reset RC DB in case an agent changes the RC url
In case a database is configured from data from a different datacenter
(or url to be completely generic) the agent is forced to reset the
database like if the version of the agent changed or the API Key
changed.
This allows users to change the site without having to manually delete
the cached configurations.