[RC] Add delay if requests for config updates are too frequent
At this moment in time, the RC backend uses the unix epoch as the
version for the TUF Director Targets file. If an agent makes multiple
requests to the backend within a second (likely due to a new client
coming online and triggering a cache bypass request) the end result is
the agent's TUF client seeing the 2nd update as the same version as the
1st, and it dropping the update silently.

We cannot change the version of the TUF Director Targets file at this
time until we get consistency across all tracer client implementations,
so we are stuck temporarily with using the unix epoch. This issue
potentially impacts clients on startup, forcing them to wait a full
agent refresh loop to get their configs when they normally would get
them much faster due to the cache bypass. To restore functionality of
the bypass with as minimal latency added as possible to guarantee we
dodge ths issue, we track the timestamp of the last update and enforce
that updates are at least a second apart.