A starting devcontainer to work with Bazel
This is a convenience for those of us working on the Linux builid from
macOS.

It's written so that Bazel outputs are written to a bind mount,
allowing the cache to survive container restarts.

Some of the shenaningans around having a bazel-user are inherited from
past experiments where running bazel as root created problems; that's
something we can revisit later.