service discovery: Enable check only if option is enabled
Currently, the Service Discovery check is always enabled but does
nothing in its Run() function if the discovery.enabled option is not
true.  This results in the service_discovery integration being listed as
active in the UI even if Service Discovery is not enabled.  To avoid
this, use the autodiscovery mechanism to enable the Service Discovery
check only if the config option is enabled.