[RC] Remove CreationTime from RC DB metadata tests
Tests involving the creation time have flaked recently due to a test
running quicker than expected for the granularity of time with which we
are testing for. We don't actually use this value in any logic for
making decisions about rebuilding the DB, it's primarily for debugging,
so rather than force this to be mockable we can just remove the
condition check from the unit tests.