Try to avoid reporting errors during system-probe startup
During startup of the agent, the checks which need system-probe start
before system-probe is up and this leads to error-level logs from the
agent and the checks temporarily being listed as broken in Fleet
Automation, and the errors reported are visibile to the user.

To improve this, first check whether system-probe is up by hitting a
generic endpoint, and if it's not, for the first few minutes, to not
report errors from the check but only emit warnings to the agent log.