usm: tests: Added utility to clean protocol map
The helper should allow us to clear the kernel mapping between connections to protocols, so
the tests won't accidentally find irrelevant connection while checking the classification.

Some tests do not allow us to determine the client socket (like kafka), so we cannot know
which client connection was used for the traffic, hence we determine the classification was
successful only by the server port. However, the client library can trigger multiple requests
to the server (like kafka does), and we can classify a connection we don't care about (for example
classifying a connection by produce request and marking the test as passed, although we wanted to
check the fetch classification). The helper allows us to clear on demand the protocol mapping in
the kernel, and to increase the confidence we have in the tests