feat(autoinstrumentation): don't mutate istio-proxy
- https://datadoghq.atlassian.net/browse/INPLAT-454
- Closes #33296

There are cases with auto-instrumentation where you want to skip entire
containers in a pod.

An alternative here would be to build container targetting into workload
selection. That feels a lot heavier and if you're running a something
like istio (something with another mutating controller) you'll need to
change every target to explicitly choose a container.

There are _known_ container names (like `istio-proxy`) that will not
be supported for auto-instrumentation (C++) in the near future and
when conflating with an issue like #33296 where the istio controller
will take a sidecar container and move to be a new kubernetes-native
"sidecar container" (https://kubernetes.io/docs/concepts/workloads/pods/sidecar-containers/).

This doesn't fix the issue _fully_ (IE there are probably other
containers where it's posssible for another mutating controller to cause
this issue) but will give people the flexibility to disable injection at
a container+cluster level.

Co-authored-by: Kyle Nusbaum <kyle@datadog.com>