usm: protocols: Don't dispatch is protocol is not supported / enabled
If a connection was classified by CNM classification (for example redis), but the protocol wasn't enabled
or is not yet supported for the dispatcher (like AMQP), we would still try to dispatch the tail call
and cache the TCP seq. This is redundant, as the tail call will not be called, and we have a redundant
entry in the map.