system-probe: Use longer timeouts for check clients
Currently, the timeout in the system-probe API is only 5 seconds.  While
this makes sense for clients which need to access system-probe as part
of some other work (e.g. for flare generation), it may be too small for
core agent checks which rely on system-probe to do most of the work.
Spuriously timing out early in this case is not helpful since the check
does not perform its main function, and system-probe actually continues
to compute whatever it had to compute in the background (since the http
handler will run to completion), except that the result is not returned
to whoever asked for it.

Increase the timeout to handle this.  During startup, to prevent waiting
for unnecessarily long time if system-probe has issues starting up, use
the shorter timeout since the stats endpoint should return quickly.