Make system-probe client test use real transport
Currently, the client tests use TCP sockets, but this is not the actual
transport used by system-probe (UDS on Linux and named pipes on
Windows).  Modify the tests to use the correct transports to match the
real scenarios closer.  For this, move and reuse some test support code
used in the flare tests.