usm: kafka: tests: Fix flakiness of tests
Since we moved to kafka 4.0 docker, we've seen an increase in test flakes.
The flake is an outcome of KRaft algorithm being less stable, and it takes
the kafka server more time to prepare the leaders for the partitions.
Since we're creating topics with 2 partitions, sometimes the topic is stored
in a partition that does not have yet a ready leader.

The fix makes 2 changes - waits for the leaders to be assigned, and then
forcing a hard metadata refresh by calling FoceMetadtaRefresh, and performing
an actual kafka call (ping) to pull the latest metadata, and ensure the client
internal state is updated with it