service discovery: Get working directory from proc
Currently, the code only gets the working directory from the PWD
environment variable, but this does not necessarily need to be present
nor up-to-date, so try to get the working directory from the cwd symlink
in proc. Unfortunately, we can't know which source is the correct one
for the command line which we're looking at (since the working directory
can be changed at anytime during execution of the processes), so we try
both when trying to resolve relative paths.