Check for nil values and handle them as viper did
When run with DD_CONF_NODETREEMODEL=enable, this fixes the tests TestDisabledNamespacesEnvVar, TestEnabledNamespacesEnvVar, and TestLibVersionsEnvVar in pkg/clusteragent/admission/mutate/autoinstrumentation

It's arguable that these tests are behaving badly, because they construct interface values with nil data in them, which is known to be a footgun in golang. However, it's quite easy to fall into this situation and debugging what's going on is tricky; better to let the library just do the right thing.