Fix heartbeat for host SBOMs
https://github.com/aquasecurity/trivy/pull/8275 changed the
generation method for cache keys, that is used by the SBOM check
to send either a heartbeat or a full SBOM. As the key is now different
for every scan, we always send the full SBOM even if it's identical
to the previous computed one.

This commit computes the key based on the collected results
so that its stays identical if the list of components hasn't changed.