service discovery: Fix race in TestPythonFromBashScript
The logic for detecting dash vs bash has a race condition since the
forked process initially has the sh command line even in the bash case
so it's mistakenly identified as the dash case and waits for a
subprocess which never appears. Combine the two Eventually loops to fix
it.