(fleet) remove useless debian preinst
This PR removes the .deb `preinst`.

Motivation: The only operation done by that script is stoping the agent. This is not useful as the previous .deb `prerm` will stop the agent as well (and it is more logic to do that there).

I checked the very first agent 6 we shipped and it already had logic to stop the agent in the `prerm`. It also had this stoping logic in the `preinst` so I assume it was just always kept without anyone really knowing why it's there.