[smp] pr-commenter: replace commit SHA with PR URL
In some cases, the `pr-commenter` service's heuristic for determining
a PR's URL returns the "wrong" URL -- the same commit might be the
`HEAD` of another PR, possibly even in a different repository. To
avoid having this service use that PR URL heuristic, this commit
supplies the PR URL directly to SMP's `pr-commenter` payload when the
PR number is available, in place of the org/repo/commit triple
currently used in those payloads.

Signed-off-by: Geoffrey M. Oxberry <geoffrey.oxberry@datadoghq.com>