Add remote caching for docker builds and use a second build context for artifacts
This includes the logic for when to invalidate the cache and when to
write to the cache.

Using a second context for artifacts allows docker to separately hash
the build context that barely changes from the artifacts themselves
which change on every pipeline, allowing more steps to be properly
cached.