dyninst/object: introduce package
This package abstracts object files for use in dyninst. It importantly
retains access to sections for use outside of just the *dwarf.Data construction
so that the same sections can be reused for location list parsing and other
operations not directly supported by debug/dwarf. It also wraps the abstractions
provided by delve's loclist package to make location lists easier to parse.

This layer can also in the future (if we deem it relevant) model the differences
between elf and mach-o (though that's possibly over-abstracting a bit).

There's some TODOs left for supporting dwarf 5 location lists that are coming
in the next Go release (see [0]).

[0]: https://github.com/golang/go/issues/26379